ui/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-WTF==0.14.2
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
requests==2.22.0
urllib3==1.25.3
uWSGI==2.0.18
Werkzeug==0.15.4
WTForms==2.2.1